secondary/requirements.txt
aiofiles==24.1.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
priority==2.0.0
protobuf==6.32.1
Quart==0.20.0
setuptools==80.9.0
Werkzeug==3.1.3
wsproto==1.2.0
aiohttp==3.12.15
pydantic==2.8.0